feat: 커스텀 훅으로 로직 병합
