+++ src/components/ui/card/FolderCard.tsx
@@ -3,9 +3,10 @@ import Image from '@/components/common/Image';
 import clsx from 'clsx';
 import { isMobile } from 'react-device-detect';
 import { motion } from 'framer-motion';
-import { useState, useRef, useCallback, useEffect } from 'react';
+
 import FolderMenuPortal from '@/utils/FolderMenuPortal';
-import type React from 'react';
+
+import { useMenuHandler } from '@/components/hooks/menuHandler';
 
 interface ICardProps {
   category: string;
@@ -17,70 +18,20 @@ const TitleText =
   'overflow-hidden font-sans font-semibold text-ellipsis whitespace-nowrap ml-1 md:text-xl text-base';
 
 const FolderCard = ({ category, images }: ICardProps) => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [menuPosition, setMenuPosition] = useState({ x: 0, y: 0 });
-  const iconRef = useRef<HTMLDivElement>(null);
-
-  const updateMenuPosition = useCallback(() => {
-    if (iconRef.current && isMenuOpen) {
-      const rect = iconRef.current.getBoundingClientRect();
-      setMenuPosition({
-        x: rect.right - 136,
-        y: rect.bottom + 4,
-      });
-    }
-  }, [isMenuOpen]);
-
-  // 리사이즈 및 스크롤 이벤트 리스너 등록
-  useEffect(() => {
-    if (!isMenuOpen) return;
-
-    // ResizeObserver를 사용하여 화면 크기 변화 감지
-    const resizeObserver = new ResizeObserver(() => {
-      updateMenuPosition();
-    });
-
-    resizeObserver.observe(document.body);
-    window.addEventListener('resize', updateMenuPosition);
-
-    return () => {
-      resizeObserver.disconnect();
-      window.removeEventListener('resize', updateMenuPosition);
-    };
-  }, [isMenuOpen, updateMenuPosition]);
-
-  // 아이콘 기준 메뉴창 열기
-  const handleIconClick = (event: React.MouseEvent) => {
-    event.stopPropagation();
-
-    if (iconRef.current) {
-      const rect = iconRef.current.getBoundingClientRect();
-      setMenuPosition({
-        x: rect.right - 136,
-        y: rect.bottom + 4,
-      });
-    }
-    setIsMenuOpen(true);
-  };
-
-  const handleMenuClose = () => {
-    setIsMenuOpen(false);
-  };
-
-  const shouldHoverEffect = isMenuOpen;
+  const { isMenuOpen, menuPosition, iconRef, isOpen, isClose } = useMenuHandler(136);
 
   return (
     <>
       <motion.div
         // 메뉴바 떠 있다면 hover상태 강제 적용 (모바일 제외)
-        animate={{ scale: shouldHoverEffect && !isMobile ? 1.03 : 1 }}
+        animate={{ scale: isMenuOpen && !isMobile ? 1.03 : 1 }}
         whileHover={{ scale: 1.03 }}
         transition={{ duration: 0.4 }}
         className={clsx(
           isMobile ? 'min-w-40 pt-2' : 'w-1/2 lg:w-1/3 xl:w-1/4 sm:mt-2 p-2 border-transparent ', // 모바일은 카드의 너비를 고정, PC는 반응형에 따라 비율 조정
           'rounded-2xl',
           !isMobile && 'hover:shadow-md hover:border-1 hover:border-gray-300',
-          shouldHoverEffect && !isMobile && 'shadow-md border-1 border-gray-300',
+          isMenuOpen && !isMobile && 'shadow-md border-1 border-gray-300',
         )}
       >
         {/**카테고리에 카드가 하나만 있으면 폴더에 하나만, 두개 있으면 1 : 1 비율... 3개까지 표시 */}
@@ -118,21 +69,21 @@ const FolderCard = ({ category, images }: ICardProps) => {
         </div>
         <div className='flex items-center justify-between pt-2'>
           <p className={TitleText}>{category}</p>
-          <div ref={iconRef} onClick={handleIconClick}>
+          <div ref={iconRef} onClick={isOpen}>
             <FolderDetailIcon
               width={24}
               height={24}
               className={clsx(
                 'hover:text-grayBg transition-colors cursor-pointer',
                 isMobile ? 'w-6 h-6' : 'sm:w-10 w-8 sm:h-10 h-8',
-                shouldHoverEffect ? 'text-grayBg' : 'text-white',
+                isMenuOpen ? 'text-grayBg' : 'text-white',
               )}
             />
           </div>
         </div>
       </motion.div>
       {/* Portal로 렌더링되는 메뉴 */}
-      <FolderMenuPortal isOpen={isMenuOpen} onClose={handleMenuClose} position={menuPosition}>
+      <FolderMenuPortal isOpen={isMenuOpen} onClose={isClose} position={menuPosition}>
         <div className='flex flex-col w-32'>
           <p className='text-left px-1 mb-2 text-[#A4A8B2] rounded text-xs'>카테고리 설정</p>
           <button className='text-left px-1 py-3 text-stone hover:bg-gray-100 rounded text-15'>
